+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Unreleased
 
+### v1.5.0-beta3 (2020-10-09)
+
+* Init the DeviceIdentifier to a fixed value in the CLI environment without setting any cookies, to prevent
+  errors if the process has already sent output.
 * Add MutexWrapper with Mock and Db (mysql) backed implementations for preventing concurrent executions
   of code.